pyproject.toml
[tool.poetry]
name = "wunder-gkeep"
version = "0.1.0"
description = "wunderlist to google keep"
authors = ["mcmhav <mcmhav@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
gkeepapi = "^0.11.7"
requests = "^2.22"

[tool.poetry.dev-dependencies]
pyyaml = "^5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
